mex_builder.py: cache Matlab vars via pickle.
Implemented a simple mechanism for caching the Matlab environment variables via
pickle. This means that Matlab will in effect only have to be started once now.
Deleting the file causes it to be recreated.

+++ site_scons/site_tools/mex_builder.py
@@ -14,7 +14,32 @@ from __future__ import with_statement # with statement in Python 2.5
 # TODO: Find out if results can be cached. As a workaround, use interactive mode
 # and use the "build" command (short form "b").
 
-import os, tempfile, subprocess as subp, sys
+import os, tempfile, subprocess as subp, sys, pickle
+
+vars_file = '.matlab_vars_cache'
+
+def load_matlab_vars(env):
+    """Load various Matlab specific variables from a cache file via the pickle
+    module.
+    """
+
+    with open( vars_file, 'r' ) as f:
+        p = pickle.Unpickler(f)
+        env['MATLAB'] = p.load()
+
+def cache_matlab_vars(matlab_vars):
+    """Store various Matlab specific variables in a cache file via the pickle
+    module.
+    """
+
+    # create the file if it doesn't exist
+    if not os.path.isfile(vars_file):
+        os.mknod(vars_file)
+
+    # store the objects in a file
+    with open(vars_file, 'w') as f:
+        p = pickle.Pickler(f)
+        p.dump(matlab_vars)
 
 def gen_matlab_env(env, **kwargs):
     """Obtain various Matlab specific variables and put them in env['MATLAB'].
@@ -22,46 +47,47 @@ def gen_matlab_env(env, **kwargs):
     # determine if the env is supposed to be a mex extension, default to False
     is_mex_ext = kwargs.get('mex', False)
 
-    tmp_file, tmp_file_name = tempfile.mkstemp()
-    # As per Python tempfile documentation, Windows doesn't support multiple
-    # processes accessing the same file, so close it immediately.
-    os.close(tmp_file)
+    if not os.path.isfile(vars_file):
+        tmp_file, tmp_file_name = tempfile.mkstemp()
+        # As per Python tempfile documentation, Windows doesn't support multiple
+        # processes accessing the same file, so close it immediately.
+        os.close(tmp_file)
 
-    # Invoke matlab, method of doing so taken from the mlabwrap setup.py.  The
-    # usage of '10' as a newline char is needed because... maybe Python
-    # universal newlines translate to newlines Matlab doesn't like? I dunno, but
-    # in the Matlab command line '\n' works, but not in this script, even with
-    # escapes or as a raw string.
-    matlab_cmd = "fid = fopen('%s', 'wt');" % tmp_file_name + \
-            r"fprintf(fid, '%s%c%s%c%s%c', mexext, 10, matlabroot, 10, computer, 10, version, 10);" + \
-            "fclose(fid);quit;"
-    cmd_line = ['matlab', '-nodesktop', '-nosplash', '-r', matlab_cmd]
-    if os.name == "nt":
-        cmd_line[-1] = '"' + cmd_line[-1] + '"'
-        cmd_line += ['-wait'] # stop Matlab from forking
+        # Invoke matlab, method of doing so taken from the mlabwrap setup.py.  The
+        # usage of '10' as a newline char is needed because... maybe Python
+        # universal newlines translate to newlines Matlab doesn't like? I dunno, but
+        # in the Matlab command line '\n' works, but not in this script, even with
+        # escapes or as a raw string.
+        matlab_cmd = "fid = fopen('%s', 'wt');" % tmp_file_name + \
+                r"fprintf(fid, '%s%c%s%c%s%c', mexext, 10, matlabroot, 10, computer, 10, version, 10);" + \
+                "fclose(fid);quit;"
+        cmd_line = ['matlab', '-nodesktop', '-nosplash', '-r', matlab_cmd]
+        if os.name == "nt":
+            cmd_line[-1] = '"' + cmd_line[-1] + '"'
+            cmd_line += ['-wait'] # stop Matlab from forking
 
-    try:
-        # output to pipe to suppress output on Unix
-        subp.check_call(cmd_line, stdout=subp.PIPE)
-    except BaseException, e:
-        # PEP 352 can't go ahead quickly enough, stupid args tuple. I want the
-        # message attribute back!
-        print >> sys.stderr, "Error:", ', '.join([repr(i) for i in e.args])
-        os.remove(tmp_file_name)
-        exit("Error calling Matlab, exiting.")
+        try:
+            # output to pipe to suppress output on Unix
+            subp.check_call(cmd_line, stdout=subp.PIPE)
+        except BaseException, e:
+            # PEP 352 can't go ahead quickly enough, stupid args tuple. I want the
+            # message attribute back!
+            print >> sys.stderr, "Error:", ', '.join([repr(i) for i in e.args])
+            os.remove(tmp_file_name)
+            exit("Error calling Matlab, exiting.")
 
-    # read lines from file and remove newline chars
-    with open(tmp_file_name) as tmp_file:
-        lines = [l.strip('\n') for l in tmp_file.readlines()]
-    os.remove(tmp_file_name)
+        # read lines from file and remove newline chars
+        with open(tmp_file_name) as tmp_file:
+            lines = [l.strip('\n') for l in tmp_file.readlines()]
+            os.remove(tmp_file_name)
 
-    matlab_root = lines[1]
-    matlab_arch = lines[2].lower()
-    matlab_ver, matlab_release  = lines[3].split()
-    if matlab_arch == 'pcwin':
-        matlab_arch = 'win32'
+        matlab_root = lines[1]
+        matlab_arch = lines[2].lower()
+        matlab_ver, matlab_release  = lines[3].split()
+        if matlab_arch == 'pcwin':
+            matlab_arch = 'win32'
 
-    env['MATLAB'] = {
+        env['MATLAB'] = {
             "MEX_EXT":  "." + lines[0],
             "ROOT":     matlab_root,
             "ARCH":     matlab_arch,
@@ -70,13 +96,19 @@ def gen_matlab_env(env, **kwargs):
             "SRC":      os.sep.join([matlab_root, 'extern', 'src']),
             "INCLUDE":  os.sep.join([matlab_root, 'extern', 'include']),
             "LIB_DIR":  [os.sep.join([matlab_root, 'bin', matlab_arch])]
-            }
+        }
 
-    if matlab_arch == 'win32':
-        env['MATLAB']['LIB_DIR'] += \
-                [os.sep.join([matlab_root, 'extern', 'lib', 'win32', 'microsoft'])]
-        # TODO: test WINDOWS_INSERT_DEF option
-        # env.Replace(WINDOWS_INSERT_DEF=True)
+        if matlab_arch == 'win32':
+            env['MATLAB']['LIB_DIR'] += \
+                    [os.sep.join([matlab_root, 'extern', 'lib', 'win32', 'microsoft'])]
+            # TODO: test WINDOWS_INSERT_DEF option
+            # env.Replace(WINDOWS_INSERT_DEF=True)
+
+        print "Caching Matlab vars..."
+        cache_matlab_vars(env['MATLAB'])
+    else:
+        print "Loading Matlab vars from cache..."
+        load_matlab_vars(env)
 
     env.Append(CPPPATH = [env['MATLAB']['INCLUDE']],
             LIBPATH = [env['MATLAB']['LIB_DIR']])
